pyproject.toml
[project]
name = "smart-to-do-api"
version = "0.1.0"
description = "A robust, human-centric To Do API."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.0",
    "motor==3.4.0",
    "beanie==1.25.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "pyjwt==2.8.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "pymongo==4.6.3",
]

[tool.uv]
dev-dependencies = [
    "pytest==8.1.1",
    "httpx==0.27.0",
    "pytest-asyncio==0.23.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "db", "models", "schemas"]
